feat(auth): add an unauthorized page
added an unauthorized page that redirect to the signin page if the user is not authenticated. The said page is a guard in the app's layout that return an client page in order to get the current path and search param in the layout route.
Maybe in the future we may need to migrate to middlewares when we will find a way to use nodemailer in next edge